requirements.txt
streamlit>=1.34.0
llama-index-core>=0.11.0
llama-index-llms-openai>=0.4.7
llama-index-embeddings-openai>=0.3.1
llama-index-vector-stores-chroma>=0.4.2
# llama-index-agent>=0.3.0  # Non disponible pour l'instant
# llama-index-tools>=0.1.0  # Non disponible pour l'instant
chromadb>=0.4.24
pypdf>=4.2.0
python-dotenv>=1.0.1
openai>=1.32.0
tqdm>=4.66.4
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0